(wip) init amm_base
